requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
firebase-admin>=6.4
python-dotenv>=1.0
gunicorn>=22.0
whitenoise>=6.6
dj-database-url>=2.1
Pillow>=10.0
